feat: implement connection handling and statement preparation for TDengine
